// src/types.ts
export interface IBanData {
  /**
   * The unique id of this ban
   */
  key: string;
  /**
   * The playersName of who was banned
   * @example "Smell of curry"
   */
  playerName: string;
  /**
   * The date in MS when this player was banned
   */
  date: number;
  /**
   * The duration in ms that this ban will be for
   * if null player is banned forever
   */
  duration?: number;
  /**
   * When this ban will expire, if null it will never expire meaning player is banned forever
   */
  expire?: number;
  /**
   * The reason for why this player was banned
   * @example "Hacking"
   */
  reason?: string;
  /**
   * Who banned this player
   * @example "Smell of curry"
   * @example "Rubedo Anti Cheat"
   */
  by?: string;
}